moved make_clickable_monster_name function to helper_functions
